requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
altgraph==0.17.4
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
attrs==23.2.0
audioread==3.0.1
av==11.0.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
contourpy==1.2.0
cryptography==42.0.5
ctranslate2==4.0.0
cycler==0.12.1
decorator==5.1.1
docopt==0.6.2
einops==0.7.0
faster-whisper==0.10.0
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
greenlet==3.0.3
huggingface-hub==0.26.2
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
julius==0.2.7
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
lightning==2.2.0.post0
lightning-utilities==0.10.1
llvmlite==0.42.0
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.17.0
openai-whisper==20231117
optuna==3.5.0
packaging==23.2
pandas==2.2.0
pefile==2023.2.7
Pillow==9.5.0
platformdirs==4.2.0
pooch==1.8.1
primePy==1.3
protobuf==4.25.3
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.21
pydub==0.25.1
Pygments==2.17.2
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.1
pyparsing==3.1.1
pyreadline3==3.4.1
python-dateutil==2.8.2
pytorch-lightning==2.2.0.post0
pytorch-metric-learning==2.4.1
pytz==2024.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.5
scikit-learn==1.4.1.post1
scipy==1.12.0
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
speechbrain==0.5.16
SQLAlchemy==2.0.27
sympy==1.12
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.3.0
tiktoken==0.6.0
tokenizers==0.20.3
torch==2.2.0
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.2.0
torchmetrics==1.3.1
tqdm==4.66.2
transformers==4.46.3
ttkbootstrap==1.10.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
